Add workaround for parcel-bundler/parcel#7157
Obvious hack, would be more palatable if it didn't pollute the tree
by switching the imports, but haven't found a solution for that and
it doesn't really matter because the prod build only gets run in CI
where we don't care. The core problem is that ParcelJS thinks its
parcelRequire will work in the worker so that it can share modules
with the main bundle - instead we trick it into thinking these are
unique deps so it bundles them directly instead. Fragile AF yadda
yadda, I just want my Workers to work.